Reshuffling a func from one file to another
